Added Operator
You can now search via the enum dropdowns, or by clicking on a key to search for that key.
Does not include key combinations, nor are the operator category enums linked to the keymap, nor do they even return results at times when they should.
However, it's a start. 
Still more a mockup than an addon at this point.