Keymap Search, Keybind Highlight, UI changes
Changed the Keymap from a property to a property search:
Blender's Preferences make it appear as though there are ~30 keymaps, some of which have 'submaps'.
In practice, ~200 keymaps sit on a flat hierarchical structure separated only by name, meaning that users would be presented with a massive list of keymaps that would frequently not have enough space to display on a 2K monitor.
By changing it to a search, users can now start typing the name of the area they wish to view, or scroll through what it now a single column list.
Should be easier to use in general.

Attempted to implement a search feature for the keymap items, but ran into some issues due to my implementation. (I'm using names, Blender is expecting idnames and iterating through every item to match up names with idnames can be costly)
The longest list is (I think) 3D View, and even it is less than half the length of the Keymap list.
I don't find it too cumbersome to navigate as it is, and may just leave it as a property unless someone feels that it would massively benefit from a search.

The keybind is now highlit. Instead of a StringProperty that justifies left and sits in a black background, it is now an Operator, set to pressed (depress=True), which centers the key name and highlights it in blue. The operator has no function, likely this will be the 'rebind' button when that feature is implemented.

Small UI Changes:
-Show/Hide is now in the same box as the keyboard, mouse etc controls, and had 'Onscreen' removed from its label
-There is now a separate button to show/hide the Numpad
-New default layout hides all except basic keyboard and mouse (was hoping this would resolve the squashed keys with default Preferences window size, but it did not)

Other Changes:
-uploaded default and expanded layout pics
-all functions that reference the keyconfig now call the active keyconfig - previously some were hardcoded to call the 'Blender' keyconfig